CamProxy/requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-decouple==3.8
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
Werkzeug==3.1.3
wsproto==1.2.0
